Made active area end where the button zone begins
This still doesn't enable resting a thumb on the trackpad most of the time. Even when it does work, the left-click is not recognized unless a finger is touching elsewhere on the trackpad, or the thumb is lifted and placed back on the pad. Most of the time resting the thumb prevents the cursor from moving.